Fixed issue with stock list and added functionality to view stock information
Stock list was previously created in a function, making inaccessible for other methods outside of that function. Fixed this by defining it outside of the function. Also, implemented the functionality to view the information from the database of the stock that the user selects by pressing the View Stock button.