Withdraw/deposit now creates monetary transactions and other changes
Withdrawing and depositing money will now make monetary transactions records. Since we decided not to give the option to delete specific transactions, I deleted the delete transaction label, button, and field. Made it so that if you search for stocks with the search field empty, it will display all stocks.