Fixed bug when attempting to buy more than total shares
Attempting to buy more than total shares would result in the creation of a null transaction that can not be deleted from within the program. This prevented users with these transactions from being able to delete their accounts. Now, when entering the amount to buy, it display a pop up error window when the value > totalShares and when the value < 1. No longer creates a null transaction.